Support for functions with vector outputs
radxeval() is now less buggy and supports functions with vector outputs.
The derivative tensor returned is in a compressed form with each column
representing derivatives of each output function.

The derivatives in a particular column are elements of the derivative
tensor stored in order of the multi-index corresponding to the
derivative (which reduces the size of the tensor from N^D to
binomial(N + D - 1, D - 1)).